pyproject.toml
[project]
name = "h-cli"
version = "0.1.0"
description = "Personal productivity CLI tool"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "types-PyYAML>=6.0.12.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/yourusername/h-cli"
Repository = "https://github.com/yourusername/h-cli.git"

[project.scripts]
h = "h.cli:app"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "."

[tool.hatch.build.targets.wheel]
packages = ["h"]
